Introduce `EmbeddedQuicChannel`
__Motivation__

Today tests mock a `QuicChannel`. Over time the mocking is becoming more verbose as we require more features in http/3, eg: while adding QPACK dynamic tables, `createStream()` are required to be implemented. As `EmbeddedChannel` exists and extending it to support `QuicStreamChannel` is not much work, it is less tedious to implement and use `EmbeddedQuicChannel` instead of the mocks.

__Modification__

- Implemented `EmbeddedQuicChannel`
- Modified all usages of mocks for the `QuicChannel` to now use `EmbeddedQuicChannel`

__Result__

Less tedious changes to tests (mocks) when implementing new http/3 features.