Set QpackAttributes in `handlerAdded()`
__Motivation__

As reported by #157; `Http3ConnectionHandler` sets `QpackAttributes` on the channel when the channel is active, but `Http3FrameCodec` requires it in `handlerAdded()` raising an exception if a stream is created before the `channelActive` is fired on the parent channel. This can happen if the stream is created from within the listener of the connect future.

__Modification__

As `QpackAttributes` do not require the channel to be active, we now create it in `handlerAdded()` unconditionally, so that it is consistent with the expectations of `Http3FrameCodec`.

__Result__

Fixes #157