QPACK: Allow stream cancellation even if the dynamic table was configured with max capacity of 0
Motivation:

It is fine to send stream cancellations even if the dynamic table had a capacity of 0 and so was disabled. We didn't follow the spec here and did case a NPE if we received a stream cancellation after the table was configured with capacity of 0.

Modifications:

- Allow stream cancellation if if the table capacity is 0
- Add unit test

Result:

Fixes https://github.com/netty/netty-incubator-codec-http3/issues/261 and https://github.com/netty/netty-incubator-codec-http3/issues/252